+++ src/help/svg.spec.ts
@@ -1,8 +1,8 @@
-import {Component} from '../component';
+import {getElement, isUpdatable} from '../public';
 
 import {svgTagNames} from './constants';
 import * as allTags from './svg';
-import {path, svg} from './svg';
+import {circle, path, svg} from './svg';
 
 test('all svg tags are defined', () => {
   expect(svgTagNames.filter((i) => !(allTags as any)[i])).toEqual(['switch']);
@@ -10,7 +10,7 @@ test('all svg tags are defined', () => {
 });
 
 test('empty svg', () => {
-  const result = path();
+  const result = path() as SVGElement;
 
   expect(result).toBeInstanceOf(SVGElement);
   expect(result.attributes.length).toBe(0);
@@ -18,7 +18,7 @@ test('empty svg', () => {
 });
 
 test('staic svg', () => {
-  const result = path({width: 100}, 'bbb');
+  const result = path({width: 100}, 'bbb') as SVGElement;
 
   expect(result).toBeInstanceOf(SVGElement);
   expect(result.attributes.length).toBe(1);
@@ -28,17 +28,16 @@ test('staic svg', () => {
 test('dynamic div', () => {
   const result = svg(() => 'bbb');
 
-  expect(result).toBeInstanceOf(Component);
+  expect(isUpdatable(result)).toBe(true);
 
-  const {element} = result;
+  const element = getElement(result);
 
   expect(element).toBeInstanceOf(SVGSVGElement);
   expect(element.attributes.length).toBe(0);
   expect(element.childNodes.length).toBe(1);
 });
 
-// test('correct typescript typings', () => {
-//   expect(svg().x).toBe('');
-//   expect(path().ariaPlaceholder).toBe('');
-//   expect(text().fontSize).toBe('');
-// });
+test('correct typescript typings', () => {
+  expect(getElement(svg({x_p: 123})).x).toBe(123);
+  expect(getElement(circle({r_p: 321})).r).toBe(321);
+});