+++ src/init.spec.ts
@@ -1,6 +1,7 @@
+import {getElement, update} from '.';
 import {init} from './init';
 import {Component} from './component';
-import {Arg, StaticArg} from './types';
+import {Arg, Fusion, StaticArg} from './types';
 
 export interface CreatorNoConf {
   <E extends Element>(element: E, args: readonly StaticArg[]): E;
@@ -55,14 +56,14 @@ test('init element with dynamic prop', () => {
   const result = init(element, [{title: () => title}]);
 
   expect(result).toBeInstanceOf(Component);
-  expect(result.element).toBe(element);
+  expect(getElement(result)).toBe(element);
 
   expect(element.attributes.length).toBe(1);
   expect(element.childNodes.length).toBe(0);
   expect(element.title).toBe('aaa');
 
   title = 'bbb';
-  result.update();
+  update(result);
 
   expect(element.attributes.length).toBe(1);
   expect(element.childNodes.length).toBe(0);
@@ -80,7 +81,7 @@ test('init element with dynamic prop override', () => {
   ]);
 
   expect(result).toBeInstanceOf(Component);
-  expect(result.element).toBe(element);
+  expect(getElement(result)).toBe(element);
 
   expect(element.attributes.length).toBe(1);
   expect(element.childNodes.length).toBe(0);
@@ -89,7 +90,7 @@ test('init element with dynamic prop override', () => {
   expect(dynamic1).toBe(112);
   expect(dynamic2).toBe(223);
 
-  result.update();
+  update(result);
 
   expect(element.id).toBe('223');
   expect(dynamic1).toBe(112); // did not increment
@@ -109,7 +110,7 @@ test('init element with event handler, it should be static', () => {
 
 test('init element with static children', () => {
   const element = document.createElement('div');
-  const result = init(element, ['one', 2, 'three']);
+  const result = init(element, ['one', 2, 'three']) as Element;
 
   expect(result).toBe(element);
   expect(result.attributes.length).toBe(0);
@@ -122,7 +123,9 @@ test('init element with static children', () => {
 
 test('init element with nested static child', () => {
   const element = document.createElement('div');
-  const result = init(element, [init(document.createElement('p'), ['Hi!'])]);
+  const result = init(element, [
+    init(document.createElement('p'), ['Hi!']),
+  ]) as Element;
 
   expect(result).toBe(element);
   expect(result.attributes.length).toBe(0);
@@ -138,7 +141,7 @@ test('init element with dynamic child', () => {
   const result = init(element, [() => child]);
 
   expect(result).toBeInstanceOf(Component);
-  expect(result.element).toBe(element);
+  expect(getElement(result)).toBe(element);
 
   expect(element.attributes.length).toBe(0);
   expect(element.childNodes.length).toBe(1);
@@ -146,7 +149,7 @@ test('init element with dynamic child', () => {
   expect(element.childNodes[0].nodeValue).toBe('123');
 
   child = 456;
-  result.update();
+  update(result);
 
   expect(element.childNodes[0].nodeValue).toBe('456');
 });
@@ -160,22 +163,22 @@ test('init element with nested dynamic child', () => {
   ]);
 
   expect(result).toBeInstanceOf(Component);
-  expect(result.element).toBe(element);
+  expect(getElement(result)).toBe(element);
 
   expect(element.attributes.length).toBe(0);
   expect(element.childNodes.length).toBe(1);
 
   expect(element.outerHTML).toBe('<div><p>1</p></div>');
 
-  result.update();
+  update(result);
 
   expect(element.outerHTML).toBe('<div><p>2</p></div>');
 
-  result.update();
+  update(result);
 
   expect(element.outerHTML).toBe('<div><p>3</p></div>');
 
-  result.update();
+  update(result);
 
   expect(element.outerHTML).toBe('<div><p>4</p></div>');
 });
@@ -187,7 +190,7 @@ describe('init element changing nested dynamic child', () => {
   const result = init(element, [() => child]);
 
   expect(result).toBeInstanceOf(Component);
-  expect(result.element).toBe(element);
+  expect(getElement(result)).toBe(element);
 
   expect(element.attributes.length).toBe(0);
   expect(element.childNodes.length).toBe(1);
@@ -213,7 +216,7 @@ describe('init element changing nested dynamic child', () => {
     ['last'],
   ])('child %p toBe %p', (val) => {
     child = val;
-    result.update();
+    update(result);
 
     if (val instanceof Element) expect(element.childNodes[0]).toBe(val);
     else if (val instanceof Component)
@@ -224,7 +227,7 @@ describe('init element changing nested dynamic child', () => {
 
 test('init element incrementing nested dynamic children', () => {
   let count = 0;
-  let cache: Component<HTMLElement>;
+  let cache: Fusion;
 
   const element = document.createElement('div');
   const result = init(element, [
@@ -233,24 +236,24 @@ test('init element incrementing nested dynamic children', () => {
     init(document.createElement('p'), [
       () =>
         cache
-          ? cache.update()
+          ? update(cache)
           : (cache = init(document.createElement('span'), [() => ++count])),
     ]),
   ]);
 
   expect(result).toBeInstanceOf(Component);
-  expect(result.element).toBe(element);
+  expect(getElement(result)).toBe(element);
 
   expect(element.attributes.length).toBe(0);
   expect(element.childNodes.length).toBe(3);
 
   expect(element.outerHTML).toBe('<div><p>1</p> <p><span>2</span></p></div>');
 
-  result.update();
+  update(result);
 
   expect(element.outerHTML).toBe('<div><p>3</p> <p><span>4</span></p></div>');
 
-  result.update();
+  update(result);
 
   expect(element.outerHTML).toBe('<div><p>5</p> <p><span>6</span></p></div>');
 });
@@ -273,7 +276,7 @@ test.each([
 
   if (typeof provided === 'function' || provided instanceof Component) {
     expect(result).toBeInstanceOf(Component);
-    expect(result.element).toBe(element);
+    expect(getElement(result)).toBe(element);
   } else {
     expect(result).toBe(element);
   }
@@ -289,7 +292,7 @@ test('init element with dynamic prop and child', () => {
   const result = init(element, [{title: () => title}, () => child]);
 
   expect(result).toBeInstanceOf(Component);
-  expect(result.element).toBe(element);
+  expect(getElement(result)).toBe(element);
 
   expect(element.attributes.length).toBe(1);
   expect(element.childNodes.length).toBe(1);
@@ -299,7 +302,7 @@ test('init element with dynamic prop and child', () => {
 
   title = 'bbb';
   child = 456;
-  result.update();
+  update(result);
 
   expect(element.attributes.length).toBe(1);
   expect(element.childNodes.length).toBe(1);
@@ -336,7 +339,7 @@ test('init element with dynamic array of children', () => {
   const result = init(element, [[111, () => 'dynamic', () => [1, 2, 3]]]);
 
   expect(result).toBeInstanceOf(Component);
-  expect(result.element).toBe(element);
+  expect(getElement(result)).toBe(element);
 
   expect(element.attributes.length).toBe(0);
   expect(element.childNodes.length).toBe(6);
@@ -348,7 +351,7 @@ test('init element with dynamic array of children', () => {
   expect(element.childNodes[4].nodeValue).toBe('3');
   expect(element.childNodes[5].nodeValue).toBe(''); // terminator
 
-  result.update();
+  update(result);
 
   expect(element.attributes.length).toBe(0);
   expect(element.childNodes.length).toBe(6);